# reference-implementation/openapi/reference-full.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PDPP Reference Implementation (full, includes /_ref)",
    "version": "0.1.0",
    "description": "Public PDPP JSON APIs plus reference-designated /_ref operator/control surfaces."
  },
...
  "paths": {
...
    "/oauth/par": {
      "post": {
        "operationId": "createPushedAuthorizationRequest",
        "tags": [
          "grants"
        ],
        "summary": "Stage a PDPP data-access request and receive a pending-consent request_uri plus authorization URL.",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Pending consent request created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "request_uri": {
                      "type": "string",
                      "pattern": "^urn:pdpp:pending-consent:"
                    },
                    "authorization_url": {
                      "type": "string",
                      "format": "uri"
                    },
                    "expires_in": {
                      "type": "integer",
                      "minimum": 1
                    }
                  },
                  "required": [
                    "request_uri",
                    "authorization_url",
                    "expires_in"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$id": "pdpp/common/PdppError",
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "type": {
                          "type": "string"
                        },
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "param": {
                          "type": "string"
                        },
                        "request_id": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "type",
                        "code",
                        "message",
                        "request_id"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Request rejected because the resolved grant contract is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$id": "pdpp/common/PdppError",
                  "type": "object",
                  "additionalProperties": false,
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": false,
                      "properties": {
                        "type": {
                          "type": "string"
                        },
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "param": {
                          "type": "string"
                        },
                        "request_id": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "type",
                        "code",
                        "message",
                        "request_id"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "properties": {
                  "client_id": {
                    "type": "string",
                    "minLength": 1
                  },
                  "client_display": {
                    "type": "object",
                    "additionalProperties": false,
                    "properties": {
                      "name": {
                        "type": "string",
                        "minLength": 1
                      },
                      "uri": {
                        "type": "string",
                        "format": "uri"
                      },
                      "logo_uri": {
                        "type": "string",
                        "format": "uri"
                      },
                      "policy_uri": {
                        "type": "string",
                        "format": "uri"
                      },
                      "tos_uri": {
                        "type": "string",
                        "format": "uri"
                      }
                    }
                  },
                  "scenario_id": {
                    "type": "string",
                    "minLength": 1
                  },
                  "authorization_details": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 1,
                    "items": {
                      "allOf": [
                        {
                          "type": "object",
                          "additionalProperties": false,
                          "properties": {
                            "type": {
                              "const": "https://pdpp.org/data-access"
                            },
                            "connector_id": {
                              "type": "string",
                              "minLength": 1
                            },
                            "provider_id": {
                              "type": "string",
                              "minLength": 1
                            },
                            "purpose_code": {
                              "type": "string",
                              "minLength": 1
                            },
                            "purpose_description": {
                              "type": "string",
                              "minLength": 1
                            },
                            "access_mode": {
                              "type": "string",
                              "enum": [
                                "single_use",
                                "continuous"
                              ]
                            },
                            "retention": {
                              "type": "object",
                              "additionalProperties": false,
                              "properties": {
                                "max_duration": {
                                  "type": "string",
                                  "minLength": 1
                                },
                                "on_expiry": {
                                  "type": "string",
                                  "minLength": 1
                                }
                              },
                              "required": [
                                "max_duration",
                                "on_expiry"
                              ]
                            },
                            "streams": {
                              "type": "array",
                              "minItems": 1,
                              "items": {
                                "type": "object",
                                "additionalProperties": false,
                                "properties": {
                                  "name": {
                                    "type": "string",
                                    "minLength": 1
                                  },
                                  "necessity": {
                                    "type": "string",
                                    "enum": [
                                      "required",
                                      "optional"
                                    ]
                                  },
                                  "view": {
                                    "type": "string",
                                    "minLength": 1
                                  },
                                  "fields": {
                                    "type": "array",
                                    "minItems": 1,
                                    "items": {
                                      "type": "string",
                                      "minLength": 1
                                    }
                                  },
                                  "time_range": {
                                    "type": "object",
                                    "additionalProperties": false,
                                    "properties": {
                                      "since": {
                                        "type": "string",
                                        "minLength": 1
                                      },
                                      "until": {
                                        "type": "string",
                                        "minLength": 1
                                      }
                                    }
                                  },
                                  "resources": {
                                    "type": "array",
                                    "items": {
                                      "type": "string",
                                      "minLength": 1
                                    }
                                  },
                                  "client_claims": {
                                    "type": "object",
                                    "additionalProperties": true
                                  }
                                },
                                "required": [
                                  "name"
                                ]
                              }
                            }
                          },
                          "required": [
                            "type",
                            "purpose_code",
                            "access_mode",
                            "streams"
                          ]
                        },
                        {
                          "oneOf": [
                            {
                              "required": [
                                "connector_id"
                              ]
                            },
                            {
                              "required": [
                                "provider_id"
                              ]
                            }
                          ]
                        }
                      ]
                    }
                  }
                },
                "required": [
                  "client_id",
                  "authorization_details"
                ]
              }
            }
          }
        }
      }
    },
...
  },
...
}